criacao classe EnderecoEntity
