+++ src/main/java/br/com/alura/comex/modelo/ClienteEntity.java
@@ -1,5 +1,6 @@
 package br.com.alura.comex.modelo;
 
+import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -16,28 +17,20 @@ public class ClienteEntity {
 	private String nome;
 	private long cpf;
 	private String telefone;
-	private String rua;
-	private String numero;
-	private String complemento;
-	private String bairro;
-	private String cidade;
-	private String estado;
-	// TODO separar endereco | @ManyToOne
+	
+	// @ManyToOne
+	@Embedded
+	private EnderecoEntity endereco;
+
+	public ClienteEntity() {
+	}
 
 	public ClienteEntity(String nome, long cpf, String telefone, String rua, String numero, String complemento,
 			String bairro, String cidade, String estado) {
 		this.nome = nome;
 		this.cpf = cpf;
 		this.telefone = telefone;
-		this.rua = rua;
-		this.numero = numero;
-		this.complemento = complemento;
-		this.bairro = bairro;
-		this.cidade = cidade;
-		this.estado = estado;
-	}
-
-	public ClienteEntity() {
+		this.endereco = new EnderecoEntity(rua, numero, complemento, bairro, cidade, estado);
 	}
 
 	public String getNome() {
@@ -65,52 +58,8 @@ public class ClienteEntity {
 		this.telefone = telefone;
 	}
 
-	public String getRua() {
-		return rua;
-	}
-
-	public void setRua(String rua) {
-		this.rua = rua;
-	}
-
-	public String getNumero() {
-		return numero;
-	}
-
-	public void setNumero(String numero) {
-		this.numero = numero;
-	}
-
-	public String getComplemento() {
-		return complemento;
-	}
-
-	public void setComplemento(String complemento) {
-		this.complemento = complemento;
-	}
-
-	public String getBairro() {
-		return bairro;
-	}
-
-	public void setBairro(String bairro) {
-		this.bairro = bairro;
-	}
-
-	public String getCidade() {
-		return cidade;
-	}
-
-	public void setCidade(String cidade) {
-		this.cidade = cidade;
-	}
-
-	public String getEstado() {
-		return estado;
-	}
-
-	public void setEstado(String estado) {
-		this.estado = estado;
+	public EnderecoEntity getEndereco() {
+		return endereco;
 	}
 
 	public long getId() {
