+++ src/main/java/br/com/alura/comex/modelo/EnderecoEntity.java
@@ -1,0 +1,51 @@
+package br.com.alura.comex.modelo;
+
+import javax.persistence.Embeddable;
+
+@Embeddable
+public class EnderecoEntity {
+
+	private String rua;
+	private String numero;
+	private String complemento;
+	private String bairro;
+	private String cidade;
+	private String estado;
+
+	public EnderecoEntity() {
+	}
+
+	public EnderecoEntity(String rua, String numero, String complemento, 
+			String bairro, String cidade, String estado) {
+		this.rua = rua;
+		this.numero = numero;
+		this.complemento = complemento;
+		this.bairro = bairro;
+		this.cidade = cidade;
+		this.estado = estado;
+	}
+
+	public String getRua() {
+		return rua;
+	}
+
+	public String getNumero() {
+		return numero;
+	}
+
+	public String getComplemento() {
+		return complemento;
+	}
+
+	public String getBairro() {
+		return bairro;
+	}
+
+	public String getCidade() {
+		return cidade;
+	}
+
+	public String getEstado() {
+		return estado;
+	}
+}
